+++ Products/CMFPlone/_scripts/compile_resources.py
@@ -17,11 +17,15 @@ package_json_contents = """{
     "grunt-contrib-requirejs": "~1.0.0",
     "grunt-contrib-uglify": "~1.0.1",
     "grunt-contrib-watch": "~1.0.0",
-    "grunt-debug-task": "~0.1.8",
     "grunt-sed": "~0.1.1",
     "less-plugin-inline-urls": "^1.1.0"
   }
 }"""
+
+if os.name == 'nt':
+    NPM_CMD = 'npm.cmd'
+else:
+    NPM_CMD = 'npm'
 
 
 def generate_package_json(base_path):
@@ -129,13 +133,19 @@ def main(argv=sys.argv):
     # generates only if not already there
     generate_package_json(base_path)
     if not args.skip_npminstall:
-        cmd = ['npm', 'install']
+        cmd = [NPM_CMD, 'install']
         print('Setup npm env')
         print('Running command: %s' % ' '.join(cmd))
         subprocess.check_call(cmd)
 
     # Generate Gruntfile
-    grunt = os.path.join(base_path, 'node_modules', 'grunt-cli', 'bin', 'grunt')
+    grunt = os.path.join(
+        base_path,
+        'node_modules',
+        'grunt-cli',
+        'bin',
+        'grunt'
+    )
     if not args.skip_gruntfile:
         generate_gruntfile(
             base_path,