+++ CHANGES.rst
@@ -10,6 +10,9 @@ Changelog
 
 Breaking changes:
 
+- Remove discontinued module ``grunt-debug-task`` from ``plone-compile-resources``.
+  [jensens]
+
 - Remove deprecated resource registrations for ``mockup-parser`` and ``mockup-registry`` from mockup-core.
   Use those from patternslib instead.
   [thet]
@@ -66,6 +69,9 @@ New features:
 Bug fixes:
 
 - Use ``Plone Test Setup`` and ``Plone Test Teardown`` from ``plone.app.robotframework`` master.  [maurits]
+
+- Let npm install work on windows for plone-compile-resources.
+  [jensens]
 
 - Don't fail, when combining bundles and the target resource files (``BUNLDE-compiled.[min.js|css]``) do not yet exist on the filesystem.
   Fixes GenericSetup failing silently on import with when a to-be-compiled bundle which exists only as registry entry is processed in the ``combine-bundle`` step.